# docs-new/api/openapi.yaml
openapi: 3.0.3
info:
  title: Limuru Cottage Hospital Queue Management API
  description: |
    REST API for the Limuru Cottage Hospital Queue Management System.
    
    This API provides endpoints for managing patient queues, appointments,
    clinical documentation, and staff communication.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Public endpoints: 30 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## Localization
    The API supports English (en) and Swahili (sw) via the Accept-Language header.
  version: 2.0.0
  contact:
    name: Limuru Cottage Hospital IT
    email: it-support@limuruhospital.co.ke
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-logo:
    url: https://limuruhospital.co.ke/logo.png
servers:
  - url: https://api.limuruhospital.co.ke
    description: Production server
  - url: https://api-staging.limuruhospital.co.ke
    description: Staging server
  - url: http://localhost:8787
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/patients:
    get:
      tags: [Patients]
      summary: List patients
      description: Get a paginated list of patients
      operationId: listPatients
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          description: Search by name, email, phone, or patient number
          schema:
            type: string
      responses:
        '200':
          description: Patient list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPatients'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Patients]
      summary: Register patient
      description: Register a new patient
      operationId: registerPatient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPatientRequest'
      responses:
        '201':
          description: Patient registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/patients/{patientId}:
    get:
      tags: [Patients]
      summary: Get patient details
      description: Get detailed information about a patient
      operationId: getPatient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          description: Patient ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetailResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Patients]
      summary: Update patient
      description: Update patient information
      operationId: updatePatient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          description: Patient ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/patients/{patientId}/history:
    get:
      tags: [Patients]
      summary: Get patient history
      description: Get visit history for a patient
      operationId: getPatientHistory
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          description: Patient ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Patient history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientHistoryResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== APPOINTMENTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid credentials"
...
    RegisterPatientRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+?[0-9]{9,15}$'
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        address:
          type: string
        nationalId:
          type: string
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        bloodType:
          type: string
          enum: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-']
        allergies:
          type: string
        medicalConditions:
          type: string
        language:
          type: string
          enum: [en, sw]
          default: en
    UpdatePatientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        address:
          type: string
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        bloodType:
          type: string
        allergies:
          type: string
        medicalConditions:
          type: string
    PatientResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Patient'
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientNumber:
          type: string
          example: "LCH-2024-0001"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        address:
          type: string
        nationalId:
          type: string
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        bloodType:
          type: string
        allergies:
          type: string
        medicalConditions:
          type: string
        language:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PatientDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            patient:
              $ref: '#/components/schemas/Patient'
            stats:
              type: object
              properties:
                totalVisits:
                  type: integer
                lastVisit:
                  type: string
                  format: date
    PaginatedPatients:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Patient'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            hasMore:
              type: boolean
    PatientHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            visits:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  ticketNumber:
                    type: string
                  department:
                    type: string
                  status:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  completedAt:
                    type: string
                    format: date-time
                  diagnosis:
                    type: string

    # ============ Appointments ============
...